+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * `[errutil]` Added method `Reset` resetting `Errors` instance
 * `[sliceutil]` Added methods `Copy`, `CopyInts` and `CopyFloats` for copying slices
+* `[csv]` Code refactoring
 * `[sliceutil]` Code refactoring
 
 ### 12.30.0