+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### 12.31.0
 
-* `[errutil]` Added method `Reset` resetting `Errors` instance
+* `[errutil]` Added method `Reset` for resetting `Errors` instance
 * `[sliceutil]` Added methods `Copy`, `CopyInts` and `CopyFloats` for copying slices
 * `[csv]` Code refactoring
 * `[sliceutil]` Code refactoring